Update to dataLoad.R
This update adds code to strip the Start and End station names to bare
site names. These sites names are then used to join to the Borough and
Coordinate data from the last update to create one wholistic dataset
which contains all of the necessary information like Borough, Latitude
and Longitude for both Start and End Stations.