Updates for adding Redistribution Indicator and Improved commenting
The changes as part of this update include calculation and addition of a
redistribution indicator. The redistribution idicator is added an
assumption that when the End station and Start station of consecutive
rows do not match that bicycle was redistributed to the Start station of
the later row.
More descriptive comments have been added for better readability.